% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_cash_flow_statement.R
\name{q_cash_flow_statement}
\alias{q_cash_flow_statement}
\title{q_cash_flow_statement}
\usage{
q_cash_flow_statement(ticker, freq = c("quarterly", "annual"))
}
\description{
q_cash_flow_statement
}
